# Conflicts:
#	app/src/main/java/freed/cam/apis/camera2/modules/RawZslModuleApi2.java
#	app/src/main/java/freed/cam/ui/themesample/handler/UserMessageHandler.java
#	build.gradle

+++ app/src/main/java/freed/gl/program/GLComputeProgram.java
@@ -8,7 +8,7 @@ public abstract class GLComputeProgram extends GLProgram implements GLComputePro
 
     private Shader computeShader;
 
-    public GLComputeProgram(int glesVersion) {
+    public GLComputeProgram(float glesVersion) {
         super(glesVersion);
     }
